requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
c2pa-python==0.27.1
cachetools==6.2.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
faiss-cpu==1.13.2
fastapi==0.135.1
fastapi-users==15.0.4
fastapi-users-db-sqlalchemy==7.0.0
filelock==3.25.2
fsspec==2026.1.0
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.3.5
hyperframe==6.1.0
idna==3.11
imagekitio==5.2.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
makefun==1.16.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypdf==1.0
networkx==3.6.1
numpy==2.3.5
openpyxl==3.1.5
orjson==3.11.7
packaging==24.2
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pinecone==8.1.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.9.4
pluggy==1.6.0
ply==3.11
postgrest==2.28.3
propcache==0.4.1
pwdlib==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.11.0
pyparsing==3.3.2
pypdf==6.6.2
pypdfium2==5.3.0
pyroaring==1.0.4
pytesseract==0.3.13
pytest==9.0.2
python-dateutil==2.9.0.post0
python-discovery==1.1.3
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
realtime==2.28.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.50.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tqdm==4.67.2
transformers==5.0.0
# Editable Git install with no remote (tutorialFastapi==0.1.0)
-e c:\users\sanjhuk chakma\desktop\projectcse\tutorialfastapi
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uv==0.10.8
uvicorn==0.41.0
virtualenv==21.2.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.23.0
zstandard==0.25.0
